Prevent downloads of existing files
Use `os.path.exists` to check whether or not the incoming file has
already been downloaded, and only download the media file if it does not
exist

+++ main.py
@@ -87,10 +87,13 @@ def main():
         destination = f"./{destination_folder_name}/{platform}"
         if not os.path.exists(destination):
             os.mkdir(destination)
-        urllib.request.urlretrieve(
-            media_url,
-            f"./{destination_folder_name}/{platform}/{file_name}.{file_type}",
+        download = (
+            f"./{destination_folder_name}/{platform}/{file_name}.{file_type}"
         )
+        if not os.path.exists(download):
+            urllib.request.urlretrieve(
+                media_url, download,
+            )
 
 
 if __name__ == "__main__":
